kyc_backend/requirements.txt
# Web API
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv>=1.0.1

# Vision / liveness / face verification
numpy==1.26.4
opencv-python-headless==4.9.0.80


# LLM reply provider
google-genai>=0.8.0
google-cloud-vision==3.7.2
sqlalchemy==2.0.30
pillow==10.4.0
mediapipe==0.10.14
face-recognition==1.3.0
# Age prediction
deepface==0.0.93
tensorflow==2.15.1
pgvector==0.4.2